Limit query interface results using search.rowid DESC
This vastly increases the speed of the query for any search with a
decent number of matches. It can be slightly less accurate since
it depends on the insert order of events, but that usually won't
be a problem for a query limited to a single network. The litterbox
query interface isn't meant to exhaustively list results either.